Consolidate modules for clearer navigation.
Merge core types and HomeKit position persistence, drop the BlindService wrapper in favor of free dispatch helpers on BlindController, and inline small HAP/homekit modules that added file hops without separate change reasons.
