+++ src/commands/serve.rs
@@ -6,7 +6,6 @@ use crate::config::ResolvedConfig;
 use crate::controller::BlindController;
 use crate::homekit;
 use crate::server::{serve, AppState};
-use crate::service::BlindService;
 
 pub async fn run(resolved_config: ResolvedConfig) -> Result<()> {
     let report = doctor::collect(&resolved_config, 0).await;
@@ -17,8 +16,7 @@ pub async fn run(resolved_config: ResolvedConfig) -> Result<()> {
 
     let controller =
         Arc::new(BlindController::with_driver(resolved_config.config.driver_config()).await?);
-    let blinds = Arc::new(BlindService::new(controller.clone()));
-    let shared_state = Arc::new(AppState::new(blinds));
+    let shared_state = Arc::new(AppState::new(controller.clone()));
 
     let hap_handles = if resolved_config.config.homekit {
         match homekit::start(controller).await {
