--- src/core/mod.rs
@@ -1,7 +1,0 @@
-//! Domain vocabulary shared across drivers, transports, and HomeKit.
-
-mod channel;
-mod command;
-
-pub use channel::Channel;
-pub use command::Command;
