--- src/homekit/config.rs
@@ -1,5 +1,0 @@
-pub const HAP_PORT: u16 = 5010;
-pub const MODEL: &str = "Somfy Telis 4";
-/// Accessory Category Identifier advertised over mDNS and encoded into setup QR payloads.
-pub const HAP_CATEGORY: &str = "2";
-pub const MDNS_NAME_PREFIX: &str = "Somfy";
